Implementation of mix-in HasDensity
